+++ src/gdconvert/conversion_array.hpp
@@ -21,6 +21,7 @@ class NDArray;
 using namespace godot;
 
 std::shared_ptr<va::VArray> ndarray_as_dtype(const NDArray& ndarray, va::DType dtype);
+std::shared_ptr<va::VArray> array_as_varray(const Array& array);
 std::shared_ptr<va::VArray> variant_as_array(const Variant& array);
 std::shared_ptr<va::VArray> variant_as_array(const Variant& array, va::DType dtype, bool copy);