+++ src/vatensor/vassign.hpp
@@ -27,6 +27,9 @@ namespace va {
 
 	std::shared_ptr<VArray> get_at_mask(const VRead& varray, const VRead& mask);
 	void set_at_mask(VWrite& varray, VRead& mask, VRead& value);
+
+	std::shared_ptr<VArray> get_at_indices(const VRead& varray, const VRead& indices);
+	void set_at_indices(VWrite& varray, VRead& indices, VRead& value);
 }
 
 #endif //NUMDOT_VASSIGN_H