+++ docs/setup/changelog.rst
@@ -15,6 +15,7 @@ Upcoming Changes (main branch)
 **Added**
 
 - Added boolean mask indexing, e.g. ``a.set(5, nd.greater(a, 5))``.
+- Added index list indexing, e.g. ``a.set(5, [[0, 1], [4, 2]])``.
 - Added ``array.copy()`` and ``nd.copy(array)`` functions.
 - Added ``array.copy()`` and ``nd.copy(array)`` functions.
 - Added ``positive`` and ``negative`` functions.